FinMem/requirements.txt
alpaca-trade-api
numpy
pandas
requests
neo4j
matplotlib
pytest
python-dotenv 